Restructured intermediary command execution to different module
