+++ src/cmd/mod.rs
@@ -1,0 +1,3 @@
+pub mod intermediary;
+
+pub use intermediary::IntermediaryCommand;
